wronakrzysztof_hw3/requirements.txt
cycler==0.9.0
matplotlib==1.5.0
numpy==1.10.1
pandas==0.17.0
pyephem==3.7.6.0
pyparsing==2.0.5
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
wheel==0.24.0
